Undo folder rename
