+++ packages/block-editor/src/components/grid-visualizer/use-get-blocks-before-current-cell.js
